Doh.  Fix a 64-bit-ism by using uintptr_t for a temporary lock variable instead of int.
git-svn-id: svn+ssh://svn.freebsd.org/base/head@67396 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ sys/kern/kern_mutex.c
@@ -232,7 +232,7 @@ mtx_enter_hard(struct mtx *m, int type, int saveintr)
 		CTR3(KTR_LOCK, "mtx_enter: 0x%p contested (lock=%p) [0x%p]",
 		    m, m->mtx_lock, RETIP(m));
 		while (!_obtain_lock(m, p)) {
-			int v;
+			uintptr_t v;
 			struct proc *p1;
 
 			mtx_enter(&sched_lock, MTX_SPIN | MTX_RLIKELY);
